Extract generatePath to shared helper _generatePath.ts
