+++ packages/next-typed-href/src/index.ts
@@ -24,6 +24,8 @@
  * $href({ route: "/users/", searchParams: { q: "hello" }, hash: "top" })
  * // => "/users/?q=hello#top"
  */
+import { generatePath } from "./_generatePath";
+
 export function defineTypedHref<
   Routes extends string,
   RouteParamsMap extends Record<Routes, Record<string, unknown>>,
@@ -46,7 +48,7 @@ export function defineTypedHref<
         }
     : never;
 
-  function generatePath<T extends Routes>(options: PathOptionsFor<T>): string {
+  function resolvePath<T extends Routes>(options: PathOptionsFor<T>): string {
     if (!("routeParams" in options)) return options.route;
 
     const { routeParams } = options as unknown as {
@@ -54,52 +56,11 @@ export function defineTypedHref<
       routeParams: Record<string, string | string[] | undefined>;
     };
 
-    return (
-      options.route
-        // Optional Catch-all Segments [[...rest]]
-        .replace(/\[\[\.\.\.(.+?)\]\]/g, (match) => {
-          const key = match.replace(/[[\].]/g, "");
-          const value = routeParams[key];
-
-          if (value === undefined) return "";
-
-          if (Array.isArray(value)) {
-            return value.map((v) => encodeURIComponent(v)).join("/");
-          }
-
-          throw new Error(`Invalid optional catch-all param: ${key} with value: ${String(value)}`);
-        })
-        // Catch-all Segments [...rest]
-        .replace(/\[\.\.\.(.+?)\]/g, (match) => {
-          const key = match.replace(/[[\].]/g, "");
-
-          if (!(key in routeParams)) throw new Error(`Missing catch-all param: ${key}`);
-          const value = routeParams[key];
-
-          if (Array.isArray(value)) {
-            return value.map((v) => encodeURIComponent(v)).join("/");
-          }
-
-          throw new Error(`Invalid catch-all param: ${key} with value: ${String(value)}`);
-        })
-        // Dynamic Segments [slug]
-        .replace(/\[(.+?)\]/g, (match) => {
-          const key = match.replace(/[[\]]/g, "");
-
-          if (!(key in routeParams)) throw new Error(`Missing route param: ${key}`);
-          const value = routeParams[key];
-
-          if (typeof value === "string") {
-            return encodeURIComponent(value);
-          }
-
-          throw new Error(`Invalid route param: ${key} with value: ${String(value)}`);
-        })
-    );
+    return generatePath(options.route, routeParams);
   }
 
   function $href<T extends Routes>(options: PathOptionsFor<T>): string {
-    const path = generatePath(options);
+    const path = resolvePath(options);
     const search = options.searchParams
       ? `?${new URLSearchParams(options.searchParams).toString()}`
       : "";
